Move models and repositories to litellm/ root
Both backend (management endpoints) and gateway routes use models and
repositories, so they should not be nested under either. This moves
litellm/backend/models/ to litellm/models/ and
litellm/gateway/repositories/ to litellm/repositories/.

The backend and gateway __init__.py files remain as backwards-compatible
re-exports from the new locations.
