--- litellm/backend/models/__init__.py
@@ -1,25 +1,0 @@
-"""
-Domain models for LiteLLM backend.
-"""
-
-from litellm.backend.models.budget import Budget
-from litellm.backend.models.credentials import Credentials
-from litellm.backend.models.model import Model
-from litellm.backend.models.object_permission import ObjectPermission
-from litellm.backend.models.organization import Organization
-from litellm.backend.models.project import Project
-from litellm.backend.models.team import Team
-from litellm.backend.models.user import User
-from litellm.backend.models.verification_token import VerificationToken
-
-__all__ = [
-    "Budget",
-    "Credentials",
-    "Model",
-    "ObjectPermission",
-    "Organization",
-    "Project",
-    "Team",
-    "User",
-    "VerificationToken",
-]
