--- litellm/gateway/repositories/__init__.py
@@ -1,31 +1,0 @@
-"""
-Repository classes for database operations.
-"""
-
-from litellm.gateway.repositories.budget_repository import BudgetRepository
-from litellm.gateway.repositories.config_repository import ConfigRepository
-from litellm.gateway.repositories.credentials_repository import CredentialsRepository
-from litellm.gateway.repositories.model_repository import ModelRepository
-from litellm.gateway.repositories.object_permission_repository import (
-    ObjectPermissionRepository,
-)
-from litellm.gateway.repositories.organization_repository import OrganizationRepository
-from litellm.gateway.repositories.project_repository import ProjectRepository
-from litellm.gateway.repositories.team_repository import TeamRepository
-from litellm.gateway.repositories.user_repository import UserRepository
-from litellm.gateway.repositories.verification_token_repository import (
-    VerificationTokenRepository,
-)
-
-__all__ = [
-    "BudgetRepository",
-    "ConfigRepository",
-    "CredentialsRepository",
-    "ModelRepository",
-    "ObjectPermissionRepository",
-    "OrganizationRepository",
-    "ProjectRepository",
-    "TeamRepository",
-    "UserRepository",
-    "VerificationTokenRepository",
-]
