+++ litellm/models/__init__.py
@@ -1,0 +1,25 @@
+"""
+Domain models for LiteLLM backend.
+"""
+
+from litellm.models.budget import Budget
+from litellm.models.credentials import Credentials
+from litellm.models.model import Model
+from litellm.models.object_permission import ObjectPermission
+from litellm.models.organization import Organization
+from litellm.models.project import Project
+from litellm.models.team import Team
+from litellm.models.user import User
+from litellm.models.verification_token import VerificationToken
+
+__all__ = [
+    "Budget",
+    "Credentials",
+    "Model",
+    "ObjectPermission",
+    "Organization",
+    "Project",
+    "Team",
+    "User",
+    "VerificationToken",
+]
