+++ litellm/models/budget.py
@@ -7,7 +7,7 @@ from typing import Any, Dict, List, Optional
 
 from pydantic import Field
 
-from litellm.backend.models.base import DomainModel
+from litellm.models.base import DomainModel
 
 
 class Budget(DomainModel):
