+++ litellm/repositories/budget_repository.py
@@ -4,8 +4,8 @@ Budget repository for database operations on LiteLLM_BudgetTable.
 
 from typing import Any, Dict, List, Optional, Type
 
-from litellm.backend.models.budget import Budget
-from litellm.gateway.repositories.base_repository import BaseRepository
+from litellm.models.budget import Budget
+from litellm.repositories.base_repository import BaseRepository
 
 
 class BudgetRepository(BaseRepository[Budget]):
